Refactored patient models subdir a bit to deal with a circular import.
