+++ lib/models/patient/main.py
@@ -7,3 +7,4 @@ from lib.models.patient.encounter_crud import *
 from lib.models.patient.encounter_model import *
 from lib.models.patient.patient_crud import *
 from lib.models.patient.patient_model import *
+from lib.models.patient.placeholders import *
