+++ lib/models/patient/patient_crud.py
@@ -5,7 +5,6 @@ from typing import Any, Dict, List, Union, cast
 
 from lib.db.surreal import AsyncDbController, DbController
 from lib.models.patient.common import PatientDict
-from lib.models.patient.encounter_crud import add_some_placeholder_encounters
 from lib.models.patient.patient_model import Patient
 from settings import logger
 
@@ -54,43 +53,6 @@ def store_patient(db: Union[DbController, AsyncDbController], patient: Patient) 
         return result
     else:
         return {}
-
-
-def add_some_placeholder_patients(db: Union[DbController, AsyncDbController]) -> None:
-    """
-    Adds some placeholder patients for testing purposes.
-    :param db: DbController instance connected to SurrealDB.
-    :return: None
-    """
-    import random
-    from datetime import datetime
-
-    # Generate 5 random patients
-    for i in range(5):
-        demographic_no = random.randint(100, 999)
-        first_name = f"FirstName{i+1}"
-        last_name = f"LastName{i+1}"
-        date_of_birth = datetime.now().replace(year=datetime.now().year - random.randint(20, 60)).isoformat()
-        location = (f"City{i+1}", f"State{i+1}", f"Country{i+1}", f"ZipCode{i+1}")
-        sex = 'r' if random.choice([True, False]) else 'm'  # Randomly assign 'r' or 'm'
-        phone = f"555-01{i+1:02d}{random.randint(1000, 9999)}"
-        email = "patient1@gmail.com"
-
-        patient = Patient(
-            demographic_no=str(demographic_no),
-            first_name=first_name,
-            last_name=last_name,
-            date_of_birth=date_of_birth,
-            location=location,
-            sex=sex,
-            phone=phone,
-            email=email
-        )
-
-        # Store the patient in the database
-        store_patient(db, patient)
-
-        add_some_placeholder_encounters(db, f"patient:{demographic_no}")
 
 
 # TODO: This is still not working 100%.
