+++ lib/routes/testing.py
@@ -6,9 +6,9 @@ from typing import Any, Dict, Tuple
 from flask import Response, jsonify, request, session
 
 from lib.db.surreal import DbController
-from lib.models.patient.main import (add_some_placeholder_patients,
-                                     create_patient, delete_patient,
-                                     get_patient_by_id, update_patient)
+from lib.models.patient.patient_crud import (create_patient, delete_patient,
+                                             get_patient_by_id, update_patient)
+from lib.models.patient.placeholders import add_some_placeholder_patients
 from settings import logger
